refactor: introduce a state snapshot to handle mutations
